+++ packages/adapter-multiversx/src/WarpMultiversxResults.ts
@@ -22,7 +22,7 @@ import {
   WarpExecution,
   WarpExecutionResults,
   WarpInitConfig,
-} from '@vleap/warps-core'
+} from '@vleap/warps'
 import { WarpMultiversxAbi } from './WarpMultiversxAbi'
 import { WarpMultiversxSerializer } from './WarpMultiversxSerializer'