+++ packages/core/tsup.config.mjs
@@ -1,4 +1,4 @@
-import { defineConfig } from 'tsup';
+import { defineConfig } from 'tsup'
 
 export default defineConfig({
   entry: ['src/index.ts'],
@@ -6,5 +6,5 @@ export default defineConfig({
   format: ['cjs', 'esm'],
   minify: true,
   clean: true,
-  external: ['@vleap/warps-core', '@vleap/warps-adapter-multiversx'],
-});
+  external: [],
+})